requirements.txt
anyio==4.4.0
appdirs==1.4.4
asgiref==3.8.1
black==24.8.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.2
h11==0.14.0
htmltools==0.5.3
idna==3.8
isort==5.13.2
Jinja2==3.1.4
linkify-it-py==2.0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdit-py-plugins==0.4.1
mdurl==0.1.2
mypy==1.11.2
mypy-extensions==1.0.0
numpy==2.1.0
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
platformdirs==4.2.2
prompt-toolkit==3.0.36
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
questionary==2.0.1
requests==2.32.3
shiny==1.0.0
six==1.16.0
sniffio==1.3.1
starlette==0.38.3
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
urllib3==2.2.2
uvicorn==0.30.6
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.0.1
